fix: local-demo 환경이 실제로 기동·seed·smoke까지 통과하도록 수정
실행하며 드러난 문제를 고쳤다.

- V31 response_status, V28 refile_count가 smallint이라 entity의 Int와 맞지 않아
  Hibernate schema validation이 실패했다. entity 쪽이 옳으므로 DDL을 integer로
  맞춘다. V31 payload_hash char(64)도 varchar(64)로 정정했다.
- cursor key ring을 환경변수로 넘기면 relaxed binding이 하이픈을 제거해
  cursor-hmac 경로에 바인딩되지 않는다. datasource와 함께 SPRING_APPLICATION_JSON
  으로 명시 전달한다.
- seed CLI는 component scan이 없으므로 seeder와 Clock을 명시적으로 등록하고
  actuator Modulith endpoint autoconfiguration도 제외한다.
- store를 saveAndFlush로 먼저 반영한다. profile row는 JDBC로 쓰여 persistence
  context를 우회하므로 FK가 보이지 않았다.
- 주문 단건 조회 응답은 생성 envelope와 달리 order를 감싸지 않는다.

실측: seed 25행 삽입 후 재실행 0행, smoke 17단계 전부 통과.

+++ src/main/resources/db/migration/V28__create_settlement_lifecycle_adjustment_and_dispute.sql
@@ -264,7 +264,7 @@ CREATE TABLE settlement_dispute (
     settlement_item_id uuid NOT NULL REFERENCES settlement_item(id),
     store_id uuid NOT NULL REFERENCES merchant_store(id),
     previous_dispute_id uuid UNIQUE REFERENCES settlement_dispute(id),
-    refile_count smallint NOT NULL DEFAULT 0 CHECK (refile_count IN (0, 1)),
+    refile_count integer NOT NULL DEFAULT 0 CHECK (refile_count IN (0, 1)),
     state varchar(24) NOT NULL
         CHECK (state IN ('FILED', 'UNDER_REVIEW', 'ACCEPTED', 'REJECTED', 'WITHDRAWN')),
     expected_adjustment_krw bigint NOT NULL,
